# apps/web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Positions Engine API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/rules/validate": {
      "post": {
        "tags": [
          "rules"
        ],
        "summary": "Rules Validate",
        "operationId": "rules_validate_rules_validate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CatalogTextRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RulesCatalogValidationResponseModel"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/rules/preview": {
      "post": {
        "tags": [
          "rules"
        ],
        "summary": "Rules Preview",
        "operationId": "rules_preview_rules_preview_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CatalogTextRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RulesCatalogPreviewResponseModel"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/rules/publish": {
      "post": {
        "tags": [
          "rules"
        ],
        "summary": "Rules Publish",
        "operationId": "rules_publish_rules_publish_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CatalogPublishRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RulesCatalogPublishResponseModel"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CatalogDiffModel": {
        "properties": {
          "added": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Added"
          },
          "removed": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Removed"
          },
          "changed": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Changed"
          }
        },
        "type": "object",
        "required": [
          "added",
          "removed",
          "changed"
        ],
        "title": "CatalogDiffModel"
      },
      "CatalogPublishRequest": {
        "properties": {
          "catalog_text": {
            "type": "string",
            "minLength": 1,
            "title": "Catalog Text"
          },
          "author": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 256
              },
              {
                "type": "null"
              }
            ],
            "title": "Author"
          }
        },
        "type": "object",
        "required": [
          "catalog_text"
        ],
        "title": "CatalogPublishRequest"
      },
      "CatalogTextRequest": {
        "properties": {
          "catalog_text": {
            "type": "string",
            "minLength": 1,
            "title": "Catalog Text"
          }
        },
        "type": "object",
        "required": [
          "catalog_text"
        ],
        "title": "CatalogTextRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "RulesCatalogPreviewResponseModel": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "counters": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Counters"
          },
          "top": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Top"
          },
          "errors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Errors"
          },
          "diff": {
            "$ref": "#/components/schemas/CatalogDiffModel"
          }
        },
        "type": "object",
        "required": [
          "ok",
          "counters",
          "top",
          "errors",
          "diff"
        ],
        "title": "RulesCatalogPreviewResponseModel"
      },
      "RulesCatalogPublishResponseModel": {
        "properties": {
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          },
          "updated_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated By"
          }
        },
        "type": "object",
        "required": [
          "version",
          "updated_at"
        ],
        "title": "RulesCatalogPublishResponseModel"
      },
...
      "RulesCatalogValidationResponseModel": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "counters": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Counters"
          },
          "top": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Top"
          },
          "errors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Errors"
          }
        },
        "type": "object",
        "required": [
          "ok",
          "counters",
          "top",
          "errors"
        ],
        "title": "RulesCatalogValidationResponseModel"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}